+++ src/Framework/Replication/Quorums/Quorum.h
@@ -1,6 +1,8 @@
 #ifndef QUORUM_H
 #define QUORUM_H
 
+#include <stdint.h>
+
 /*
 ===============================================================================
 
@@ -18,10 +20,10 @@ public:
 	virtual ~Quorum() {}
 
 	virtual unsigned		GetNumNodes() const = 0;
-	virtual const unsigned*	GetNodes() const = 0;
+	virtual const uint64_t*	GetNodes() const = 0;
 	
-	virtual void			RegisterAccepted(unsigned nodeID) = 0;
-	virtual void			RegisterRejected(unsigned nodeID) = 0;
+	virtual void			RegisterAccepted(uint64_t nodeID) = 0;
+	virtual void			RegisterRejected(uint64_t nodeID) = 0;
 	virtual void			Reset() = 0;
 
 	virtual bool			IsRoundRejected() const = 0;
